Change file name: `index.tsx` -> `Carousel.tsx`
