+++ src/components/index.tsx
@@ -1,1 +1,1 @@
-export { default as Carousel } from './Carousel';
+export * from './Carousel';
